fix: expand negated vector comparisons
Logical negation of a vector comparison should expand into a disjunction of the negations of elementwise comparisons. This PR fixes this behavior by adding an expression visitor for type and shape inference, and a DNF transformer that makes use of the inferred shapes to expand logical negations of vector comparisons as necessary. The Canonical form transformer is modified to use this DNF transformer as its base. The test suite has been updated to account for the new Canonical form behavior due to these expansions, and hundreds of unit tests have been added for the behavior of the inference visitor and DNF transformer. Two additional system tests have also been added to ensure correct operation.

Refs: #69